+++ test/index.test.ts
@@ -1,4 +1,3 @@
-import { autorun } from 'mobx';
 import { Normi } from '../src';
 
 test('normal merge', () => {
@@ -112,86 +111,4 @@ test('custom id function', () => {
   expect(val.value.first).toEqual('tuna');
   expect(val.value.instrument).toEqual('flugelhorn');
   expect(appleColin.value.username).toEqual('colin');
-});
-
-test('update node + child node + unrelated data', () => {
-  const normi = new Normi();
-
-  const mainId = '3d16368b-79cc-48f4-8b72-e514ec99315c';
-  const childId = '7c22ad2a-641a-11eb-ae93-0242ac130002';
-  const unrelatedId = '85b4abfe-641a-11eb-ae93-0242ac130002';
-
-  normi.merge({
-    id: mainId,
-    title: 'first post',
-  });
-
-  let updates: any[] = [];
-  const fn = jest.fn(() => {
-    const data = JSON.parse(JSON.stringify(normi.nodes[mainId]));
-    updates.push(data);
-  });
-  autorun(fn);
-
-  // update main id
-  normi.merge({
-    id: mainId,
-    foo: 'first post',
-    author: {
-      id: childId,
-      name: 'colin',
-    },
-  });
-  // update child
-  normi.merge({
-    id: childId,
-    github: 'https://github.com/colinhacks',
-  });
-
-  // update unrelated data
-  normi.merge({
-    id: unrelatedId,
-    foo: 'bar',
-  });
-
-  const numCalls = fn.mock.calls.length;
-  if (numCalls !== 3) {
-    console.error(`Exepected 3 calls`);
-    console.log('Data', updates);
-  }
-
-  expect(fn).toHaveBeenCalledTimes(3);
-  expect(updates).toMatchInlineSnapshot(`
-    Array [
-      Object {
-        "value": Object {
-          "id": "3d16368b-79cc-48f4-8b72-e514ec99315c",
-          "title": "first post",
-        },
-      },
-      Object {
-        "value": Object {
-          "author": Object {
-            "id": "7c22ad2a-641a-11eb-ae93-0242ac130002",
-            "name": "colin",
-          },
-          "foo": "first post",
-          "id": "3d16368b-79cc-48f4-8b72-e514ec99315c",
-          "title": "first post",
-        },
-      },
-      Object {
-        "value": Object {
-          "author": Object {
-            "github": "https://github.com/colinhacks",
-            "id": "7c22ad2a-641a-11eb-ae93-0242ac130002",
-            "name": "colin",
-          },
-          "foo": "first post",
-          "id": "3d16368b-79cc-48f4-8b72-e514ec99315c",
-          "title": "first post",
-        },
-      },
-    ]
-  `);
 });
